fix: upgraded all libs except jest

+++ scripts/cvtool_delete_retention.sh
@@ -1,12 +1,12 @@
 #!/usr/bin/env sh
 . /secret/.env
 
-RETENTION_DATE="$(date --date="120 days ago" +'%Y-%m-%d')"
+RETENTION_DATE="$(date --date='120 days ago' +'%Y-%m-%d')"
 VOLUME_CVTOOL_BACKUP=/var/lib/docker/volumes/root_cvtool_backup/_data
 VOLUME_MONGODB_BACKUP=/var/lib/docker/volumes/root_mongodb_backup/_data
 
-rm -rf ${VOLUME_CVTOOL_BACKUP}/${RETENTION_DATE}
-rm -rf ${VOLUME_MONGODB_BACKUP}/${RETENTION_DATE}
+rm -rf "${VOLUME_CVTOOL_BACKUP:?}/${RETENTION_DATE}"
+rm -rf "${VOLUME_MONGODB_BACKUP:?}/${RETENTION_DATE}"
 
 lftp -c " \
   set ftp:ssl-allow true \
